8.1. Publish State Change Validator Logic
We're working on adding some pretty complex validation rules around who can publish or unpublish a CheeseListing. To get this logic right, what we really need to know is how the $isPublished field is changing... like, is it changing from false to true? Or true to false? Or maybe it's not changing at all because the PUT request is only updating other fields.
And hey! We already know how to get the original data from Doctrine! We did it in CheeseListingDataPersister
Oh, and by the way: if your API resource is not an entity - which is totally allowed and something that we'll talk about later - then you can get the original object by injecting the RequestStack, getting the current request and then reading the previous_data attribute.
If that attribute is not there, then you know that your object is being created.

Fetching the Original Entity Data
Ok: let's get the original data just like we did before. Add a public function __construct() with EntityManagerInterface . I'll do my normal trick of hitting Alt+Enter and going to "Initialize properties" to create that property and set it
Below in the method, we can say  = $this->entityManager->getUnitOfWork()->getOriginalEntityData($value). Let's dd($originalData) to make sure it's working
Spin over and try the test:

 symfony php bin/phpunit --filter=testPublishCheeseListingValidation
Got it! It's the same array that we saw earlier and it correctly shows that isPublished is originally false!