8.2. Publish State Change Validator Logic
Remove the dd() and we also don't need this null check
You do need that when you add a validator to a property - which might be null - but not when your constraint is above a class. We will always get a CheeseListing object.
For the previous is published value, let's say: $previousIsPublished = $originalData['isPublished'] ?? false in case it's not set. That's the same thing we did in CheeseListingDataPersister.
Let's start by checking to see if the published field has not changed: if $previousIsPublished === $value->getIsPublished(), then we don't need to do any special validation. Just return, and I'll add a comment.

Not Allowing Short Descriptions
The first case in our test is that an owner cannot publish if the description is less than 100 characters.
To enforce this add if $value->getIsPublished(). Inside, we know that the field is changing from false to true: this listing is being published. So if strlen($value->getDescription()) < 100, we need to fail! Copy the buildViolation() code from below and move it up here.
The argument is the validation message... and this is coming from the ValidIsPublished annotation.
This allows you to customize the message when you use it via annotation options:

 219 lines  src/Entity/CheeseListing.php
/**
 * @ValidIsPublished()
 */
class CheeseListing
{
}
But... I'm not going to use that: we're not creating a reusable constraint... so it's simpler to keep everything in one spot:

 Cannot publish: description is too short!

We also don't need a parameter: that's if you need a dynamic wildcard in your message. Oh, but I will add ->atPath('description'). That's nice: it will make the validation failure look like it's attached to the description field even though we added the constraint to the entire class.
If the length is long enough, just return.
Testing time! Run the test and...

  symfony php bin/phpunit --filter=testPublishCheeseListingValidation
It still fails... but... yes! It's now failing on "admin can publish a short description". Over in CheeseListingResourceTest, our first assertion passed! It is returning a 400 status code