8.3. Publish State Change Validator Logic
Allowing Admin Users to Publish
It was failing before because when we log in as an admin user, it was also returning a 400 status code instead of allowing this. To fix that, we need to see if the user is an admin. Add a second argument to the constructor Security $security. I'll initialize this property.
Then below, update the if statement: if the description is too short and not $this->security->isGranted('ROLE_ADMIN'), fail validation. I'll add a comment to summarize this
This is true test-driven development. Try the test again:

  symfony php bin/phpunit --filter=testPublishCheeseListingValidation
It fails... but we are further!

  Normal user cannot unpublish

If we look at the test... yea! The second and third cases are now passing and we're down to the unpublish logic. You can start to see why breaking each test case into its own method might be a bit cleaner, even if it's more work up-front.