8.4. Publish State Change Validator Logic
Adding the Unpublish Validation
Let's add the unpublishing validation logic, which is a bit easier: only an admin can unpublish. At the bottom of the validator - thanks to the return statement and the fact that we checked to make sure that the isPublished is changing, at the bottom, we know that this cheese listing is being un published.
If not $this->security->isGranted('ROLE_ADMIN') - if we're not an admin, then this is not allowed. Copy $context->buildViolation() from earlier, give it a nice message:

 Only admin users can unpublish

And remove the atPath(): this has nothing to do with the description
Try it!

 symfony php bin/phpunit --filter=testPublishCheeseListingValidation
And... yes! That huge test now passes. Thanks to our validator and the original data, we're able to write the exact logic we need.

400 vs 403 Errors
When we set this up, we chose to use 400 validation errors:

 199 lines  tests/Functional/CheeseListingResourceTest.php
class CheeseListingResourceTest extends CustomApiTestCase
{
    public function testPublishCheeseListingValidation()
    {
        $this->assertResponseStatusCodeSame(400, 'description is too short');
        $this->assertResponseStatusCodeSame(400, 'normal user cannot unpublish');
    }
}
Which is really nice because the user gets a 400 status code and can see a collection of descriptive validation errors.
If you wanted, you could instead return a 403 access denied, which might especially make sense when a normal user tries to unpublish. How could we do that?
One of the really cool things about the way Symfony is architected is that we are free - at any point during the request - to throw an AccessDeniedException. So literally, in the middle of the validator, we can say throw new AccessDeniedException() - make sure you to get the one from the Security component. I'll say:

 Only admin users can unpublish

To see this in action, run the test again:

 symfony php bin/phpunit --filter=testPublishCheeseListingValidation
This will fail but... sweet! We can see the response: a giant 403 error.