8.5. Publish State Change Validator Logic
I'll comment that out and keep my validation logic. We now have a RESTful way to publish a listing and execute custom logic! In the fourth part of this series, we'll talk about other ways that we could have accomplished this, like using the Messenger integration or creating a truly custom operation. But I really like this solution.

Next: let's talk about how to add a completely custom field to your resource: a field that doesn't live in your entity and that might even require a service to calculate. We actually did this in a previous tutorial... but it was so custom that it didn't show up in our documentation. Let's learn an even better way.