9.1. Completely Custom Field via a Data Provider
Most of the time, the fields in our API match the properties in our entity. But we know that doesn't always need to be the case. In the previous tutorial, there were several times when we needed to add a custom field. For example, in CheeseListing, we added a shortDescription field to the API by adding a getShortDescription() method and putting it in the cheese:read group:

 219 lines  src/Entity/CheeseListing.php
class CheeseListing
{
    /**
     * @Groups("cheese:read")
     */
    public function getShortDescription(): ?string
    {
        if (strlen($this->description) < 40) {
            return $this->description;
        }
        return substr($this->description, 0, 40).'...';
    }
}
We did the same with setTextDescription(): this added a textDescription field that could be written in the API:

 219 lines  src/Entity/CheeseListing.php
class CheeseListing
{
    /**
     * The description of the cheese as raw text.
     *
     * @Groups({"cheese:write", "user:write"})
     * @SerializedName("description")
     */
    public function setTextDescription(string $description): self
    {
        $this->description = nl2br($description);
        return $this;
    }
}
Custom Fields that Require a Service
But you can only take this so far. If you need a service to calculate the value of a custom field, then you can't use this trick because these classes don't have access to services.
We even tackled this in the previous tutorial! We added an isMe boolean field to User. We did this by leveraging a custom normalizer: src/Serializer/Normalizer/UserNormalizer:

 71 lines  src/Serializer/Normalizer/UserNormalizer.php
class UserNormalizer implements ContextAwareNormalizerInterface, CacheableSupportsMethodInterface, NormalizerAwareInterface
{
    public function normalize($object, $format = null, array $context = array()): array
    {
        $data['isMe'] = $isOwner;
    }
}
We basically added an extra field right before the array is turned into JSON.

But this solution had a flaw. Look at the documentation on the User resource: I'll open the "get" operation and, down here, you can see the schema of what's going to be returned. It does not mention isMe! This field would be there if you tried the endpoint, but our docs have no idea it exists.
And... maybe you don't care! Not every app needs to have perfect docs. But if we do care? How could we add a custom field that's properly documented?
There are a few ways, like creating a completely custom API resource class or using an output DTO. We'll talk about both of these later. But in most situations, there is a simpler solution.
In UserResourceTest, if you find testGetUser(), we do have a test that looks for the isMe field. It's false in the first test and true after we log in as this user:

 79 lines  tests/Functional/UserResourceTest.php
class UserResourceTest extends CustomApiTestCase
{
    public function testGetUser()
    {
        $client = self::createClient();
        $user = UserFactory::new()->create(['phoneNumber' => '555.123.4567']);
        $authenticatedUser = UserFactory::new()->create();
        $this->logIn($client, $authenticatedUser);
        $client->request('GET', '/api/users/'.$user->getId());
        $this->assertJsonContains([
            'username' => $user->getUsername(),
        ]);
        $data = $client->getResponse()->toArray();
        $this->assertArrayNotHasKey('phoneNumber', $data);
        $this->assertJsonContains([
            'isMe' => false,
        ]);
        // refresh the user & elevate
        $user->refresh();
        $user->setRoles(['ROLE_ADMIN']);
        $user->save();
        $this->logIn($client, $user);
        $client->request('GET', '/api/users/'.$user->getId());
        $this->assertJsonContains([
            'phoneNumber' => '555.123.4567',
            'isMe' => true,
        ]);
    }
}
Copy that method name... and let's just make sure it's passing:

symfony php bin/phpunit --filter=testGetUser
Right now... it does pass. Time to break things! In UserNormalizer, remove the isMe property. Actually, we can just return $this->normalizer->normalize() directly
This class still adds a custom group, but it no longer adds a custom field. Try the test now:

 symfony php bin/phpunit --filter=testGetUser
A lovely failure!