10.2. Leveraging the Doctrine Data Provider
To fix this, open config/services.yaml and, at the bottom, override the user data provider service: App\DataProvider\UserDataProvider. Add bind and, in this case, we're going to bind the $collectionDataProvider argument. Set this to @, then go copy the service id that we know we need, and paste.
Now it should work. Refresh the browser and... got it! Let's check the number of results and... yes! It stops at 30 users but says that there are 51 total. Pagination is back!
Now that we have a working UserDataProvider, we are ready to add any custom fields we need. Let's do it next!