11.1. Adding & Populating the Custom Field
Let's ignore all of this data provider stuff for a minute and just pretend that we want to add a nice, normal field to our API. Simple! In User, add a new property - private $isMe - and put it in the user:read group.
The only difference between this property and the other properties is that I'm not going to add @ORM\Column because I don't want to store this field in the database.
Down at the bottom of this class, go to "Code"->"Generate" - or Command+N on a Mac - and select getters and setters for the isMe field. Oh, but let's improve these: add the boolean type-hint on the argument and a bool return type
So... that's actually all we need to make this part of our API! Open a new tab and go to /api to check out the documentation. On the get endpoint, click to look at the schema. There it is! An isMe boolean field.
To be even cooler - which is always my goal - back up on the property... there it is... we can add more docs:

 Returns true if this is the currently-authenticated user
That's nice because API Platform will automatically use this in the docs. When we look at the schema now on the user operation... there it is!
In a minute, we're going to set this field in UserDataProvider. But we do have a, sort of, strange situation, because - if we ever called the isMe field outside of an API call where the data provider is called... the isMe field won't be set!
Let's be extra cautious. Down in the getter, if $this->isMe is null, it means it simply hasn't been set. Throw a new LogicException:

 The isMe field has not been initialized.