11.2. Adding & Populating the Custom Field
Setting Custom Data in the Data Provider
Let's finally set this field in UserDataProvider. My guess is that the getCollection() method will return an array of users... but let's actually check that. Add $users =, dd() and, at the bottom, return $users.
Back at the browser, find the original tab and refresh. Oh! It's not an array of users! It's a Paginator object with a Doctrine Paginator inside! So... this obviously isn't an array, but the Paginator object is iterable: we can loop over it like an array and then update the isMe field on each item.
Above this line, let's add some documentation to help my editor: I'll advertise that $users is an array of User objects... which is actually a lie... but when we loop over it, we will get User objects.