11.3. Adding & Populating the Custom Field
Now, do that loop: foreach ($users as $user). And inside say $user->setIsMe() - yay for auto-completion - and set this to true to start.
Let's see if it shows up! Move over, refresh and... yes! Every record has isMe: true.

Setting isMe Based on the Current User
Setting this to the correct value is probably the easiest part of the whole process. Start by adding a second argument to the constructor - Security $security - so we can get the authenticated user. I'll hit Alt+Enter and go to Initialize Properties to create that property and set it.
Now, before the loop, set $currentUser = $this->security->getUser() and set isMe with a simple $currentUser === $user.
I love it! Try this one last time and... nice! The first one has isMe: true and then isMe is false for the others.
We did it! Oh, except that this only works for the collection endpoint. Try going to /api/users/1.jsonld. Yep!

 The isMe field has not been initialized

That makes sense: we only added the logic to the collection data provider, not the item data provider, which is still being done by the core Doctrine item provider. If we want to output the field here, we need to do a little bit more work.
Let's do that next and also find one other - kind of surprising - spot where we also need to set this field.