12.1. Custom Item Data Provider
We've successfully used our collection data provider to populate the isMe field on each User resource returned by the collection operation
If we go to /api/users.jsonld, we see it!

But if we go to an individual item, we get this error... because the isMe field is not being set yet. We can also see this in our tests - the testGetUser() method looks for the isMe field on an item endpoint. Try it:

  symfony php bin/phpunit --filter=testGetUser
And this should fail with the same error as the browser. Hmm, it did fail... but not quite like I expected. Ah, I need to make my test smarter. Before checking the JSON, add $this->assertResponseStatusCodeSame(200)
Now when we try the test:

  symfony php bin/phpunit --filter=testGetUser
Yes! It was failing with the same 500 error... and now it's more obvious.

Anyways, to make this test pass, we also need to populate the isMe field when a single User is fetched. Yep, we have a collection data provider, now we need an item data provider.

You can do this in a separate class, but it's totally fine to put both providers in the same file.