12.2. Custom Item Data Provider
Creating the Item Data Provider
Add a third interface called - wait for this mouthful -DenormalizedIdentifiersAwareItemDataProviderInterface:
Wow. If you jump into this class, it extends a less strict ItemDataProviderInterface. I'm implementing that other crazy interface because that's what the core Doctrine item data provider uses... and I want to be able to pass it the same arguments.
Ok: go to the "Code"->"Generate" menu - or Command+N on a Mac - and implement the one method we need: getItem()
Then let's immediately inject the core Doctrine item provider. Add the new argument: ItemDataProviderInterface $itemDataProvider. Hit Alt+Enter to initialize that property and then... because I'm a bit obsessive about order, I'll make sure this is the second property
Finally, down in getItem(), return $this->itemDataProvider->getItem() - yes, I totally just forgot the itemDataProvider part, I'll catch that in a second - and pass this $resourceClass, $id, $operationName and also $context.
To tell Symfony to specifically pass us the Doctrine item provider, go back to services.yaml: we need to add one more bind. The name of the argument is $itemDataProvider - so I'll copy that - and that should be passed the same service id as above, but with the name "item". If you used debug:container, that's what you would see.
Beautiful! Before we try this, let's jump straight in and set the isMe field. To do that, instead of returning, add $item = and I'll - as usual - put some phpdoc above this because we know this will be a User object or null.
It will be null if the id wasn't found in the database. To prevent that being a problem, if not, $item, then, return null.
At this point, we know we have a User object. So we can say $item->setIsMe() $this->security->getUser() === $item. Finish by returning  at the bottom.
Ok! We're ready to try the test:

 symfony php bin/phpunit --filter=testGetUser

Come on Ryan!

Now when we try it:

symfony php bin/phpunit --filter=testGetUser
Green! Congrats team! We just added a custom isMe field that is returned on both the item and collection operations and is properly documented.
But... surprise! There is one last spot where the data is missing! Open UserResourceTest and go to the top: the first test is about creating a user. Copy that method name and run the tests with --filter=testCreateUser:

 symfony php bin/phpunit --filter=testCreateUser
Ah! A 500 error! Apparently the isMe field in this User object is never set!

And... if you think about it... that makes sense! This is the one situation where the data provider system is never called! It's called when you load a collection of items or a single item, but when you're creating a new item, there was never anything to load!