12.3. Custom Item Data Provider
Don't Forget Populating in the Data Persister
But... surprise! There is one last spot where the data is missing! Open UserResourceTest and go to the top: the first test is about creating a user. Copy that method name and run the tests with --filter=testCreateUser:

 symfony php bin/phpunit --filter=testCreateUser
Ah! A 500 error! Apparently the isMe field in this User object is never set!

And... if you think about it... that makes sense! This was the one situation where the data provider system is never called! It's called when you load a collection of items or a single item, but when you're creating a new item, there was never anything to load!
The fix for this is to also make sure that we set this field in the data persister. At the top, add one more argument - Security $security - and then initialize that property.
Below, in persist() - we could add the logic in the if statement where we know this is a new item - but it doesn't hurt to set it every time a User is saved. Add $data->setIsMe($this->security->getUser() === $data)
Try the test now:

symfony php bin/phpunit --filter=testCreateUser
We got it!

What we just did is a really natural way to use entities in API Platform... but also have the flexibility to add custom fields that require a service to set their value.
But, I admit, it's not the easiest thing. This required both a collection and item data provider... and we also needed to add the same code in a data persister.
What I love about this solution is that the field isn't really custom: it's a real field that we add to our ApiResource class. The tricky part is populating that field. And it turns out, there are several other ways to do this, which might be better depending on the situation. Let's talk about those next.