13.1. Setting a Custom Field Via a Listener
When you need to add a custom field... and you need a service to populate the data on that field, you have 3 possible solutions: you can create a totally custom API resource class that's not an entity, you can create an output DTO or you can do what we did: add a non-persisted field to your entity. I like this last option because it's the least... nuclear. If most of your fields come from normal persisted properties on your entity, creating a custom resource is overkill and output DTO's - which are really cool - come with a few drawbacks.
So that's what we did: we created a non-persisted, normal property on our entity, exposed in our API. And populated it in a data provider.
But in reality, there are multiple ways that we could set that field. The data provider solution is the pure API Platform solution. The downside is that if you use your User object in some code that runs outside of an API Platform API call, the  field won't be set!
That might be ok... or you might not even have that situation. But let's look at another idea. What if we create a normal, boring Symfony event listener that's executed early during the request and we set the  field from there.
Let's try it! First, remove our current solution: in UserDataPersister I'll comment-out the $data->setIsMe() and add a comment that this will now be set in a listener.
Then over in UserDataProvider, I'll do the same thing with the first setIsMe()... and the second

Sweet! We are back to the broken state where the  field is never set.

Creating the Event Subscriber
To create the event listener, find your terminal and run:

 php bin/console make:subscriber
Well, this will really create an event subscriber, which I like a bit better. Let's call it SetIsMeOnCurrentUserSubscriber. And for the event, we want kernel.request. Well, that's its old name. Its new name is this RequestEvent class. Copy that and paste below.
Perfect! Let's go check out the new class in src/EventSubscriber/. And... brilliant! The onRequestEvent() will now be called when the RequestEvent is dispatched, which is early in Symfony