13.2. Setting a Custom Field Via a Listener
Populating the Field
So our job is fairly simple! We need to find the authenticated User if there is one, and if there is, call setIsMe(true) on it.
Add public function __construct() with a Security $security argument. I'll hit Alt+Enter and go to "Initialize properties" to create that property and set it.
Then down in onRequestEvent(), start with: if not $event->isMasterRequest(), then return.
That's not super important, but if your app uses sub-requests, there's no reason for this code to also run for those. If you don't know what I'm talking about and want to, check out our Symfony Deep Dive Tutorial.
Anyways, get the user with $user = $this->security->getUser() and add some PHPDoc above this to help my editor: we know this will be a User object or null if the user isn't logged in.
If there is no user, just return. But if there is a user, call $user->setIsMe(true)
Cool! We just set the  field on the authenticated User object. One cool thing about Doctrine is that if API platform later queries for that same user, Doctrine will return this exact object in memory, which means that the  field will be set to true.
We're now setting the $isMe field for the current user, but purposely not setting it for all other User objects. In the User class, let's now default $isMe to false to mean:
  Hey! If we did not set this, it must mean that this is not the currently-authenticated user.

Down in getIsMe(), the LogicException is no longer needed. Testing time! At your browser, refresh the item endpoint and... got it! And if we go to /api/cheeses.jsonld... the first item has isMe: true and the others are isMe: false. Love it!
But... what if the object that we need to set the data on is not the User object? How could we get access to the "current API resource object" from inside of our listener? And what if it's a collection endpoint? How could we get access to all the objects that are about to be serialized? Let's chat about the core API Platform event listeners next.