15.1. Doctrine postLoad Listener
Let's create another custom field on User. I know, we're getting crazy, but I want to keep seeing how far we can push API Platform's abilities.
This new field will be similar to isMe... but just different enough that it will lead us to an alternate solution. Suppose some users are MVPs... but being an MVP isn't as simple as having a boolean isMvp column in the database. Nope, we need to use a super fancy calculation that requires a service to determine if a User is an MVP.
So here's our goal: expose a new isMvp field to our API whose value will be populated via custom logic in a service.

Creating the isMvp Property
To start, you know the drill, we're going to add this as a real property in our entity but not persist it to Doctrine. Call it  and let's add some documentation that can be used by the API docs:
  Returns true if this user is an MVP

Next, down at the bottom, go to "Code"->"Generate" - or Command+N on a Mac - and generate the getter and setter. But rename the method to getIsMvp().
The property-access component does support looking for "isser" methods... but for that to work, the property would have to be called just mvp to look for an isMvp() method. Because our property is called isMvp, we need a getIsMvp() method.
Anyways, now if we move over and go to /api, we're hoping that the new field will already be in the docs. If you look under the user endpoint and go to schema... boom! An isMvp boolean field.