15.2. Doctrine postLoad Listener
Testing the Behavior
Of course, that field will always be false because we're never setting it. That's our next job. Well, let's start by describing the behavior in a test.
In our app, a user will be an MVP if their username - which is a field on User - contains the word "cheese".
Okay, in reality, we could accomplish this with just a custom getter method that looks for cheese in the  property and returns true or false. We don't actually need a service... but let's pretend we do.
In UserResourceTest, scroll down to testGetUser().
When we create this user, we pass it a phoneNumber, but the rest of the fields come from the UserFactory class - from the getDefaults() method.
Yep, it's using a random username from Faker.
Now, let's be a bit more specific: pass username set to cheesehead.
A few lines later, we make a request to GET this user, assert that the response is 200 and check for the username in assertJsonContains(). Let's also check for isMvp set to true.
That's good enough for now! Copy the method name and run that test:

  symfony php bin/phpunit --filter=testGetUser
And... it fails! Woo!